+++ noa-ui/browser/urls.py
@@ -2,7 +2,7 @@ from django.urls import path
 from . import views
 
 urlpatterns = [
-    path("", views.map_view, name="map_view"),
+    path("", views.results, name="results"),
     # path("search/", views.search, name="search"),
     path('results/', views.results, name='results'),  # Results view
     path('submit-order/', views.submit_order, name='submit_order'),