+++ noa-preprocess/CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.3.0] - 2025-03-27
+### Fixed
+- Fixed tile search in QI files bug (#122)
 ### Changed
 - Store clipped in different folder. Remove clipped prefix (#47)